requirements.txt
streamlit
pandas
numpy
opencv-python-headless==4.8.1.78
pydub==0.25.1
streamlit_webrtc==0.47.1
nfcpy
flask==1.0.2
pyrebase4
streamlit_option_menu